+++ hutool-poi/src/test/java/cn/hutool/poi/excel/test/ExcelWriteTest.java
@@ -450,15 +450,26 @@ public class ExcelWriteTest {
 			rows.add(tempList);
 		}
 		ExcelWriter writer = ExcelUtil.getWriter("D:\\test\\multiSheet.xlsx", "正常数据");
+		writer.addHeaderAlias("1", "row1");
+		writer.addHeaderAlias("3", "row2");
+		writer.setOnlyAlias(true);
+
 		writer.write(rows, true);
 		writer.autoSizeColumnAll();
 
+		//表2
 		writer.setSheet("当前重复数据");
+		writer.clearHeaderAlias();
+		writer.addHeaderAlias("3", "行3");
+		writer.addHeaderAlias("1", "行1");
 		writer.write(rows, true);
 		writer.autoSizeColumnAll();
+
+		//表3
 		writer.setSheet("历史重复数据");
 		writer.write(rows, true);
 		writer.autoSizeColumnAll();
+
 		writer.close();
 	}